# openapi.yaml
openapi: 3.1.0
info:
  title: lil-fleet API
  version: 1.0.0
  description: Manifest-locked lifecycle control for Docker-hosted inference models.
servers:
  - url: http://127.0.0.1:8090
security:
  - bearerAuth: []
paths:
...
  /v1/gpus:
    get:
      operationId: listGPUs
      description: >-
        GPU inventory from nvidia-smi joined with the manifest topology and Fleet's current GPU
        assignments. The snapshot is cached for about 5 seconds and refreshed by at most one
        nvidia-smi run at a time (5 s timeout). If nvidia-smi fails, the last snapshot is served
        with stale=true while it is under 60 seconds old; otherwise the endpoint returns 503.
      responses:
        "200":
          description: GPU inventory
          content:
            application/json:
              schema: {$ref: "#/components/schemas/GPUInventory"}
        "503": {description: nvidia-smi failed and no recent snapshot exists (code gpu_query_failed), content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
...
  /v1/deployments/switch:
    post:
      operationId: switchModel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [model_id]
              properties:
                model_id: {type: string}
      responses:
        "200": {description: Model was already ready}
        "202": {$ref: "#/components/responses/OperationAccepted"}
        "400": {description: Invalid body or attempted override, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
...
  responses:
    OperationAccepted:
      description: Lifecycle operation accepted
      headers:
        Location:
          schema: {type: string}
      content:
        application/json:
          schema:
            type: object
            required: [changed, operation]
            properties:
              changed: {type: boolean}
              operation: {$ref: "#/components/schemas/Operation"}
              warnings:
                type: array
                items: {type: string}
                description: Non-fatal notes about an explicit gpus selection, such as a pick spanning PCIe groups. Omitted when empty.
    NotFound:
      description: Model or operation not found
      content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}
    Conflict:
      description: >-
        A conflicting lifecycle operation is active (code operation_in_progress: another operation
        for this model, or any operation when concurrent_deployments is off), or insufficient
        topology-compatible GPUs/ports are free (code insufficient_resources).
      content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}
  schemas:
...
    GPUInventory:
      type: object
      required: [sampled_at, groups, gpus]
      properties:
        sampled_at: {type: string, format: date-time, description: When the nvidia-smi snapshot was taken.}
        stale: {type: boolean, description: Present and true when nvidia-smi failed and the last good snapshot (under 60 s old) is served.}
        groups:
          type: array
          description: Manifest PCIe topology groups; empty when none is configured.
          items: {type: array, items: {type: integer, minimum: 0}}
        gpus:
          type: array
          items: {$ref: "#/components/schemas/GPU"}
    GPU:
      type: object
      required: [index, name, memory_total_mib, memory_used_mib, memory_free_mib, utilization_percent, group, assigned]
      properties:
        index: {type: integer, minimum: 0}
        name: {type: string}
        memory_total_mib: {type: integer, minimum: 0}
        memory_used_mib: {type: integer, minimum: 0}
        memory_free_mib: {type: integer, minimum: 0}
        utilization_percent: {type: [integer, "null"], minimum: 0, maximum: 100, description: Null when nvidia-smi reports it unavailable.}
        group: {type: [integer, "null"], minimum: 0, description: Index into groups, or null when the GPU is in no group.}
        assigned:
          type: array
          description: Fleet deployment instances holding this GPU, including ones still loading or stopping; empty when free.
          items:
            type: object
            required: [model_id, instance_id]
            properties:
              model_id: {type: string}
              instance_id: {type: string}
...
    Operation:
      type: object
      required: [id, kind, model_id, state, created_at]
      properties:
        id: {type: string}
        kind: {type: string, enum: [activate, unload]}
        model_id: {type: string}
        instances:
          type: integer
          minimum: 0
          description: Target instance count for this lifecycle operation; unload uses zero.
        gpus:
          type: array
          items: {type: integer, minimum: 0}
          description: Explicit GPU indices requested by the load, in request order; omitted for automatic placement.
        warnings:
          type: array
          items: {type: string}
          description: Non-fatal notes about the explicit GPU selection.
        state: {type: string, enum: [pending, running, succeeded, failed]}
        error: {type: string, maxLength: 1024, description: Sanitized and redacted failure text.}
        created_at: {type: string, format: date-time}
        started_at: {type: string, format: date-time}
        finished_at: {type: string, format: date-time}
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string}
            message: {type: string}
        details: {type: object, additionalProperties: true}